+++ app/(public)/(with-sidebar)/class/exchange-posts/[id]/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import Link from "next/link";
+import { useRouter } from "next/navigation";
 import styled from "styled-components";
 
 type PageProps = {
@@ -9,15 +10,39 @@ type PageProps = {
   };
 };
 
-export default function ExchangePostDetailPage({ params }: PageProps) {
+const proposals = [
+  {
+    id: 1,
+    className: "개나리반",
+    lessonDate: "00.00.00",
+    writer: "최양진",
+    content: "내용내용내용내용내용내용내용내용내용내용내용내용",
+    createdAt: "00.00.00",
+  },
+  {
+    id: 2,
+    className: "개나리반",
+    lessonDate: "00.00.00",
+    writer: "최양진",
+    content: "내용내용내용내용내용내용내용내용내용내용내용내용",
+    createdAt: "00.00.00",
+  },
+];
+
+export default function ExchangePostAcceptedPage({ params }: PageProps) {
   const { postId } = params;
+  const router = useRouter();
+
+  const handleAcceptProposal = () => {
+    router.push(`/class/exchange-posts/${postId}/accepted`);
+  };
 
   return (
     <PageWrapper>
       <TopButtonRow>
         <ActionButton type="button">삭제</ActionButton>
         <ActionButton type="button">수정</ActionButton>
-        <LinkButton href="/exchange-posts">목록</LinkButton>
+        <LinkButton href="/class/exchange-posts">목록</LinkButton>
       </TopButtonRow>
 
       <Section>
@@ -45,22 +70,54 @@ export default function ExchangePostDetailPage({ params }: PageProps) {
       </Section>
 
       <Section>
+        <Label>만료일</Label>
+        <ValueBoxSmall>00.00.00</ValueBoxSmall>
+      </Section>
+
+      <Section>
         <Label>신청 현황</Label>
         <StatusBox>대기 중</StatusBox>
       </Section>
 
-      <Section>
-        <Label>교환 대상</Label>
+      <Divider />
 
-        <TargetMetaRow>
-          <MetaBox>• 개나리반</MetaBox>
-          <MetaBox>• 00.00.00</MetaBox>
-          <MetaBox>• 최양진</MetaBox>
-        </TargetMetaRow>
+      <ProposalHeader>
+        <ProposalTitle>교환 제안서</ProposalTitle>
+        <ActionButton type="button">작성 완료</ActionButton>
+      </ProposalHeader>
 
-        <TargetContent>내용내용내용내용내용내용내용내용내용내용내용내용</TargetContent>
-        <TargetDate>00.00.00</TargetDate>
-      </Section>
+      <FormRow>
+        <Input placeholder="반 이름" />
+        <Input placeholder="수업 일자" />
+        <Input placeholder="작성자" />
+      </FormRow>
+
+      <Textarea placeholder="내용" />
+
+      <ProposalList>
+        {proposals.map((proposal) => (
+          <ProposalCard key={proposal.id}>
+            <MetaRow>
+              <MetaBox>• {proposal.className}</MetaBox>
+              <MetaBox>• {proposal.lessonDate}</MetaBox>
+              <MetaBox>• {proposal.writer}</MetaBox>
+            </MetaRow>
+
+            <ContentBox>{proposal.content}</ContentBox>
+
+            <CardBottom>
+              <DateText>{proposal.createdAt}</DateText>
+              <ActionButton type="button" onClick={handleAcceptProposal}>
+                제안 수락하기
+              </ActionButton>
+            </CardBottom>
+          </ProposalCard>
+        ))}
+      </ProposalList>
+
+      <BottomRow>
+        <LinkButton href={`/exchange-posts/${postId}`}>기본 상세로 돌아가기</LinkButton>
+      </BottomRow>
     </PageWrapper>
   );
 }
@@ -80,7 +137,7 @@ const TopButtonRow = styled.div`
 `;
 
 const ActionButton = styled.button`
-  width: 80px;
+  min-width: 80px;
   height: 48px;
   border: none;
   background: #e6e6e6;
@@ -113,6 +170,15 @@ const ValueBox = styled.div`
   min-height: 48px;
   padding: 14px 16px;
   background: #f3f3f3;
+`;
+
+const ValueBoxSmall = styled.div`
+  width: 120px;
+  height: 48px;
+  padding: 0 12px;
+  background: #f3f3f3;
+  display: flex;
+  align-items: center;
 `;
 
 const Row = styled.div`
@@ -152,10 +218,64 @@ const StatusBox = styled.div`
   background: #f3f3f3;
 `;
 
-const TargetMetaRow = styled.div`
+const Divider = styled.hr`
+  border: none;
+  border-top: 1px solid #d7d7d7;
+  margin: 32px 0 28px;
+`;
+
+const ProposalHeader = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 20px;
+`;
+
+const ProposalTitle = styled.h2`
+  margin: 0;
+  font-size: 32px;
+  font-weight: 700;
+`;
+
+const FormRow = styled.div`
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
-  gap: 14px;
+  gap: 12px;
+  margin-bottom: 12px;
+`;
+
+const Input = styled.input`
+  height: 48px;
+  padding: 0 14px;
+  border: 1px solid #999;
+  outline: none;
+`;
+
+const Textarea = styled.textarea`
+  width: 100%;
+  min-height: 120px;
+  padding: 14px;
+  border: 1px solid #999;
+  resize: none;
+  outline: none;
+  margin-bottom: 24px;
+`;
+
+const ProposalList = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 28px;
+`;
+
+const ProposalCard = styled.article`
+  border-top: 1px solid #d7d7d7;
+  padding-top: 18px;
+`;
+
+const MetaRow = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  gap: 12px;
   margin-bottom: 12px;
 `;
 
@@ -167,14 +287,20 @@ const MetaBox = styled.div`
   align-items: center;
 `;
 
-const TargetContent = styled.div`
+const ContentBox = styled.div`
   min-height: 48px;
   padding: 14px;
   background: #f3f3f3;
 `;
 
-const TargetDate = styled.div`
-  margin-top: 10px;
+const CardBottom = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+  margin-top: 12px;
+`;
+
+const DateText = styled.span`
   color: #b8b8b8;
 `;
